requirements.txt
requests~=2.31.0
requests-oauthlib~=1.3.0
redis~=4.6.0
flask~=1.1.4
beautifulsoup4~=4.12.2
openai~=0.27.8
python-dotenv~=1.0.0
markupsafe==2.0.1
